+++ pkg/controllers/dnset/resource.go
@@ -15,6 +15,8 @@
 package dnset
 
 import (
+	"bytes"
+	"fmt"
 	"github.com/matrixorigin/matrixone-operator/api/core/v1alpha1"
 	"github.com/matrixorigin/matrixone-operator/pkg/controllers/common"
 	"github.com/matrixorigin/matrixone-operator/pkg/utils"
@@ -23,7 +25,53 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"text/template"
 )
+
+type model struct {
+	HAKeeperPort   int
+	DNServicePort  int
+	ConfigFilePath string
+}
+
+var startScriptTpl = template.Must(template.New("dnservice-start-script").Parse(`
+#!/bin/sh
+set -eu
+
+POD_NAME=${POD_NAME:-$HOSTNAME}
+ADDR="${POD_NAME}.${HEADLESS_SERVICE_NAME}.${NAMESPACE}.svc"
+ORDINAL=${POD_NAME##*-}
+UUID=$(printf '00000000-0000-0000-0000-%012x' ${ORDINAL})
+conf=$(mktemp)
+
+bc=$(mktemp)
+cat <<EOF > ${bc}
+uuid = "${UUID}"
+service-address = "${ADDR}:{{ .DNServicePort }}"
+listen-address = "0.0.0.0:{{ .DNServicePort }}"
+EOF
+
+# there is a chance that the dns is not yet added to kubedns and the
+# server will crash, wait before myself to be resolvable
+elapseTime=0
+period=1
+threshold=30
+while true; do
+    sleep ${period}
+    elapseTime=$(( elapseTime+period ))
+    if [[ ${elapseTime} -ge ${threshold} ]]; then
+        echo "waiting for dns resolvable timeout" >&2 && exit 1
+    fi
+    if nslookup ${ADDR} 2>/dev/null; then
+        break
+    else
+        echo "waiting pod dns name ${ADDR} resolvable" >&2
+    fi
+done
+
+echo "/mo-service -cfg ${conf}"
+exec /mo-service -cfg ${conf}
+`))
 
 // buildHeadlessSvc build the initial headless service object for the given dnset
 func buildHeadlessSvc(dn *v1alpha1.DNSet) *corev1.Service {
@@ -157,4 +205,41 @@ func syncPodSpec(ds *v1alpha1.DNSet, cs *kruise.CloneSet) {
 
 	ds.Spec.Overlay.OverlayPodSpec(&podSpec)
 	cs.Spec.Template.Spec = podSpec
+}
+
+// buildDNSetConfigMap return dn set configmap
+func buildDNSetConfigMap(dn *v1alpha1.DNSet) (*corev1.ConfigMap, error) {
+	conf := dn.Spec.Config
+
+	if conf == nil {
+		conf = v1alpha1.NewTomlConfig(map[string]interface{}{})
+	}
+
+	// detail: https://github.com/matrixorigin/matrixone/blob/main/pkg/cnservice/types.go
+	conf.Set([]string{"service-type"}, ServiceTypeDN)
+	conf.Set([]string{"dn", "Txn", "Storage"}, getStorageConfig(dn))
+	conf.Set([]string{"fileservice"}, getLocalStorageConfig())
+	conf.Set([]string{"fileservice"}, getSharedStorageConfig())
+	conf.Set([]string{"dn", "hakeeper-client", "service-addresses"}, getHaKeeperAdds(dn))
+	s, err := conf.ToString()
+	if err != nil {
+		return nil, err
+	}
+
+	buff := new(bytes.Buffer)
+	err = startScriptTpl.Execute(buff, &model{
+		ConfigFilePath: fmt.Sprintf("%s/%s", configPath, ConfigFile),
+	})
+
+	return &corev1.ConfigMap{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: dn.Namespace,
+			Name:      utils.GetName(dn),
+			Labels:    common.SubResourceLabels(dn),
+		},
+		Data: map[string]string{
+			ConfigFile: s,
+			Entrypoint: buff.String(),
+		},
+	}, nil
 }
